atmega_hal

Module port

Port

Example

Complete example source code can be found in the repository: atmega2560-blink.rs

let dp = atmega_hal::Peripherals::take().unwrap();
let pins = atmega_hal::pins!(dp);

let mut led = pins.pb7.into_output();

loop {
    led.toggle();
    delay_ms(1000);
}

Modules

mode

GPIO pin modes

Structs

Dynamic

PB0

PB1

PB2

PB3

PB4

PB5

PB6

PB7

PC0

PC1

PC2

PC3

PC4

PC5

PC6

PD0

PD1

PD2

PD3

PD4

PD5

PD6

PD7

Pins

Enums

DynamicPort

Traits

PinMode

PinOps

Type Aliases

Pin

Type-alias for a pin type which can represent any concrete pin.