Trait atmega_hal::spi::SpiOps

pub trait SpiOps<H, SCLK, MOSI, MISO, CS> {
    fn raw_setup(&mut self, settings: &Settings);
    fn raw_release(&mut self);
    fn raw_check_iflag(&self) -> bool;
    fn raw_read(&self) -> u8;
    fn raw_write(&mut self, byte: u8);
}

Internal trait for low-level SPI peripherals

This trait defines the common interface for all SPI peripheral variants. It is used as an intermediate abstraction ontop of which the Spi API is built. Prefer using the Spi API instead of this trait.

Required Methods

fn raw_setup(&mut self, settings: &Settings)

fn raw_release(&mut self)

fn raw_check_iflag(&self) -> bool

fn raw_read(&self) -> u8

fn raw_write(&mut self, byte: u8)