Trait atmega_hal::wdt::WdtOps

pub trait WdtOps<H> {
    type MCUSR;
    pub fn raw_init(&mut self, m: &Self::MCUSR);

    pub fn raw_start(&mut self, timeout: Timeout) -> Result<(), ()>;

    pub fn raw_feed(&mut self);

    pub fn raw_stop(&mut self);
}

Internal trait for low-level watchdog operations.

HAL users should use the Wdt type instead.

Associated Types

type MCUSR

Loading content...
Required methods

pub fn raw_init(&mut self, m: &Self::MCUSR)

Initialize the watchdog timer.

Warning: This is a low-level method and should not be called directly from user code.

pub fn raw_start(&mut self, timeout: Timeout) -> Result<(), ()>

Start the watchdog timer with the specified timeout.

If the timeout value is not supported, Err(()) should be returned.

Warning: This is a low-level method and should not be called directly from user code.

pub fn raw_feed(&mut self)

Feed this watchdog, to reset its period.

Warning: This is a low-level method and should not be called directly from user code.

pub fn raw_stop(&mut self)

Disable/stop this watchdog again.

Warning: This is a low-level method and should not be called directly from user code.

Loading content...