Crate atmega_hal

atmega-hal

Common HAL (hardware abstraction layer) for ATmega* microcontrollers.

Note: This version of the documentation was built for ATmega328P. This means that only items which are available for this MCU are visible. If you are using a different chip, try building the documentation locally with:

cargo doc --features <your-mcu> --open

Re-exports

pub use adc::Adc;

pub use i2c::I2c;

pub use spi::Spi;

pub use port::Pins;

pub use usart::Usart;

pub use wdt::Wdt;

pub use eeprom::Eeprom;

Modules

adc

Analog-to-Digital Converter

clock

Core clock speed management

delay

Delay implementations

eeprom

i2c

pac

Reexport of atmega328p from avr-device

port

prelude

Prelude containing all HAL traits

simple_pwm

spi

usart

wdt

Macros

pins

Structs

Atmega

Peripherals

All the peripherals.