Trait atmega_hal::eeprom::EepromOps

pub trait EepromOps<H> {
    const CAPACITY: u16;

    // Required methods
    fn raw_read_byte(&self, address: u16) -> u8;
    fn raw_write_byte(&mut self, address: u16, data: u8);
    fn raw_erase_byte(&mut self, address: u16);
}

Internal trait for low-level EEPROM peripherals.

This trait defines the common interface for all EEPROM peripheral variants.

Required Associated Constants

const CAPACITY: u16

Required Methods

fn raw_read_byte(&self, address: u16) -> u8

Read a single byte from offset address. Does not do a bounds check.

Warning: This is a low-level method and should not be called directly from user code.

fn raw_write_byte(&mut self, address: u16, data: u8)

Erase and write a single byte at offset address. Does not do a bounds check.

Warning: This is a low-level method and should not be called directly from user code.

fn raw_erase_byte(&mut self, address: u16)

Erase a single byte at offset address. Does not do a bounds check.

Warning: This is a low-level method and should not be called directly from user code.

Object Safety

This trait is not object safe.