Trait atmega_hal::wdt::WdtOps

pub trait WdtOps<H> {
    type MCUSR;

    // Required methods
    fn raw_init(&mut self, m: &Self::MCUSR);
    fn raw_start(&mut self, timeout: Timeout) -> Result<(), ()>;
    fn raw_feed(&mut self);
    fn raw_stop(&mut self);
}

Internal trait for low-level watchdog operations.

HAL users should use the Wdt type instead.

Required Associated Types

type MCUSR

Required Methods

fn raw_init(&mut self, m: &Self::MCUSR)

Initialize the watchdog timer.

Warning: This is a low-level method and should not be called directly from user code.

fn raw_start(&mut self, timeout: Timeout) -> Result<(), ()>

Start the watchdog timer with the specified timeout.

If the timeout value is not supported, Err(()) should be returned.

Warning: This is a low-level method and should not be called directly from user code.

fn raw_feed(&mut self)

Feed this watchdog, to reset its period.

Warning: This is a low-level method and should not be called directly from user code.

fn raw_stop(&mut self)

Disable/stop this watchdog again.

Warning: This is a low-level method and should not be called directly from user code.